Disabled mouse_move by default
This is because SquidDev-CC/CC-Tweaked#434 is still in progress. I don't
want to ship CraftOS-PC with features that aren't complete by default to
avoid causing some compatibility problems once the feature is complete.
The mouse_move event will still be available as it was in previous
builds, but it will require setting `mouse_move_throttle` to a positive
value to be enabled.