Added config load error message
Previously, any errors when reading the global config would cause a crash,
and errors when loading the per-computer config would stop
the computer.

Now, errors will be caught properly, displayed to the user, and the
session will use the default config, disallowing saving to prevent
overwriting all of the old settings.

This should hopefully be a bit more graceful than a hard stop on error.