Added preferredHardwareDriver, redid VS dependencies
CraftOS-PC now uses vcpkg for all libraries, so NuGet packages and manual
compilation is no longer required.
Windows builds will now also dynamically link with VCRT, meaning the C++
redistributable is required (it was already installed anyway).

preferredHardwareDriver selects the driver to use with the hardware renderer.
This can be one of direct3d, direct3d11, metal, opengl, opengles, opengles2,
or software. Some of these may not be supported on your system. A way to
check which are available will be available soon.