Brand-new icon!
CraftOS-PC used to use the same icon as ComputerCraft. This was because I
was lazy. The new icon is designed to be a neuomorphic twist on the original
ComputerCraft logo, while keeping the same colors and terminal.

Yes, this change is inspired in part by macOS 11's new icon scheme.
I've been meaning to do this for a while, but I never really got around to it
(like coming up with a new name :P).