Fixed ABI compatibility with unmodified Lua
The pcall patch that was used for #1 ended up breaking ABI compatibility
with libraries compiled for base Lua because it shuffled around some
basic definitions in the lua_State structure. This has been fixed, and
now Lua libraries should be loadable as plugins without (most) errors.

Also, the plugin version has been updated to 4 since this breaks
older versions of the Lua library.