Finally moved to Lua 5.2 bit32
One final bug relating to number of arguments finally made me make this change.
I should have done this a long time ago, but my own stubbornness and will to
keep maintaining my own version of the library prevented me from doing so.
Now it's not my problem if something goes wrong with the bit library - it's the
reference implementation, so it's not my problem!