Optimized scroll and clear routines
Instead of going throw row::operator= for each row modification, we're
using memmove on the underlying data source to try to optimize things.
This is likely the first step in removing the use of vector2d in terminals, moving
to a standard flat vector (or maybe even just a buffer) for all operations.