Completely reorganized code structure
This is the first step of #141. All API code has been placed in its own folder, and
no non-API code is present in those files. The externally-available code that used to
be in those files are now under three new files, roughly divided by where they used
to be defined. In addition, main structures (such as Computer) are now relocated to
a global api directory, for use by plugins. These files have been scrubbed of any references
to functions that would have to be defined in a source file. Finally, all non-exported code
has been given a static qualifier to reduce any chances of collisions.

The next step is to factor out as many extern statements as possible. When writing a
large portion of this code, I used externs as quick tape to import functions from other
source files to reduce duplication. A lot of these functions can be used by other
stuff too, so I've moved those into one of the new common headers, done during the
staticization part. However, there's still a bunch of externs left throughout the code,
mostly relating to things defined in main.cpp. I've added a template main.hpp, and I'll
be filling it in to get rid of those externs.

Also: I'm 100% sure CI will fail. Expect a long string of "Fixed CI" commits after this.
In addition, the ccemux plugin (and the plugin API in general) is broken - don't expect
that to work right now.