Fixed most warnings from ReSharper C++
This is step three of #141. I've gone through the entire project and resolved as many warnings
as I could that ReSharper C++ gave me. This should help keep the code style up-to-date
with the best C++ semantics; however, the sheer abundance of explicit C casts is likely
going to negatively affect code readability. (But it fixes the warnings!!!)

The next (and possibly final) step is to factor out whatever remaining C-style stuff
remains - especially C structs and dynamic C strings. I've already worked C strings
out of the event functions, but now to fix it for the rest of the codebase. (The HTTP API
is one of the files that has a *lot* of C code.)