Removed vestigial C-style code
This is step four, and quite possibly the last step, of #141. I've removed a bunch of old
C-style idioms left over from craftos-native, when most of the code was pure C.
This includes `typedef struct`, some C strings (especially dynamically-allocated ones), and
a few other things. In addition, the CRAFTOSPC_INTERNAL macro definitions were
removed, as they are now obsolete with the introduction of separate API headers.