Added virtual file mounts (closes #152)
The addVirtualMount function can be used to mount a FileEntry (brought from standalone)
to a path. The mounted directory will be read-only, but otherwise fully accessible as if it was
a normal file system mount. Virtual mounts are only accessible from plugins.
Virtual mounting can be used to add custom programs and APIs to the ROM inside a plugin.

To accomodate this, the IO library has been soft-replaced by CC:T's Lua implementation.
While the C version still exists, it is not recommended to use this as it will fail to open
files located on a virtual mount. Use the FS API or the Lua IO implementation instead.

In standalone mounts, the ROM and debug mounts are now also virtual mounts.

See the CCEmuX plugin for an example of virtual (multi-)mounting in use.

Also fixed CI.