Fixed WASM build, improved performance of HW pixel
The hardware renderer now uses a single streaming texture to render
pixels, instead of making a new texture for every frame. This should
hopefully improve performance and memory usage.

This also fixes some compile errors on WASM. I'm still wrangling with
Emscripten, trying to get it to build on the latest version. I've found
that the issue was introduced in 2.0.1, so I'll be releasing with 2.0.0
for now. (Provided more things don't break.)