Added pixel size to term.getSize as an option
This is a safe compromise for allowing access to the graphics mode size without breaking
compatibility. You can pass a mode like in setGraphicsMode, and it'll return the size
associated with the mode (it's essentially a binary test anyway). Default is mode false/0 to
avoid breaking CC.

This may be a breaking change anyway if someone previously put an argument in
term.getSize... but why would they? It was a 0-arg function before.