Mitigate load non-resumability with threads (closes #184)
This uses threads to allow `load` to yield without having to unwind the stack. More details on
how this works are at Computer.cpp:209. Since this does pose a potential performance issue
and possibility of memory leaks, this is behind standards mode for now. This is mostly
untested, but it does work - can't guarantee how well though.