Added automatic crash log uploading
This is only implemented in Windows for now. It searches for CraftOS-PC
crash dumps in the user's dump directory and automatically uploads
them to a private S3 bucket, where I can debug the issues sent there.
This is disabled by default for privacy, but the user is prompted about
this on first boot. I might expand snooperEnabled to add more analytics,
however. (I do love my __anonymized__ data!)

Note that this will not function unless you have an API key, which I'm
not distributing, and will be ciphered in release executables. I also
haven't tested this *at all*, as I currently don't have access to my
Windows dev machine. I give it an 85% chance of not working first try.