Fixed first frame of GIFs being darker
Something weird was going on in the palette detection code that caused
the first frame to have darker colors than the rest of the image. To
mitigate that, I've added some code to make it use the terminal's
palette instead of trying to find one itself, which is good since we
already know all of the colors that are going to appear. This should
also improve the performance of encoding since it doesn't have to go
through the entire image to find the palette. I also reduced the bit
depth of palette indices, which should reduce file size a tiny bit.