Added single window mode
Single window mode uses one output window to render all windows. This is
an extension of CLI mode's behavior to all other renderers. For the GUI
renderer, use Ctrl+Alt+Left/Right to switch windows (Cmd+Option+L/R on
Mac). CLI mode retains the same switching options as before. Raw
mode and TRoR do not have ways to switch windows yet, but this will be
available in the future.

This change is a prerequisite to adding multi-window support on mobile platforms.