Unicode String Support Test
This branch is a test for the possibility of Unicode support in future
versions of ComputerCraft.

Unicode strings are stored in a new UTFString type. This type is a
wrapper around a C++ UTF-32 string, with the same methods
as the base string library. UTFStrings can be created by calling
`UTFString` with a string, plus an optional second argument to
specify whether the string should be interpreted as raw UTF-8 data.

A UTFString is mostly compatible with normal strings, and a VM
extension allows them to be compared with strings. They also
have a version of the string library available, but note that they are
*not* compatible with `string` itself. You will need to replace direct
`string` calls with method calls, e.g. `string.sub(str, i)` -> `str:sub(i)`.

To support showing Unicode characters on-screen, the terminal has
been adjusted to use UTF-32 characters, and characters outside the
normal CC-ANSI range will be rendered with a (bundled) monospace Unicode
font. (This may be changed in the future to use system or other fonts.)
Unicode input is accepted as well by sending a UTFString as the second
parameter to the char event.

The `term` and `fs` APIs have been updated to support UTFStrings.
Passing a UTFString to `term` is the only way to write Unicode text to
the screen - normal strings keep the same behavior.
If you pass a UTFString to `fs` functions, the path will be converted to
UTF-8 when accessing the filesystem. There is also a new mode modifier,
'u', that will cause read methods to return a UTFString instead of a
normal string. This allows reading Unicode text safely. 'u' is mutually
exclusive with 'b', and is redundant with write modes because normal
text-write file handles accept both types of strings.

On the Lua side, some functions have been updated so far. `print`,
`cc.expect`, and `cc.pretty` support them pretty well, and `edit` has
been slightly adjusted but it doesn't work very well right now.
Otherwise, it remains pretty untested.