Rewrote a good chunk of speaker.playAudio
This now uses a queue system instead of direct playback. This fixes some
clipping issues due to the end of every block being set to 0 (since
SDL2_mixer has static chunk sizes, so any audio that is not an
even multiple of that size must be padded). This technically goes against
how CC handles it, but if I don't do this there will be terrible clipping.