Added DAP debugger frontend (WIP)
This adds support for the Debug Adapter Protocol, used by VS Code for
debugger extensions. It allows CraftOS-PC to function as a debugger in
VS Code with a basic extension wrapper around the executable.
This is still a work in progress, as the actual server is still incomplete
and missing numerous commands, but it works for very basic usage.