Fixed #327 (hopefully), closed #337
Also fixed keys repeating after holding Ctrl+R/S/T.

The WebSocket server API now uses a new function, `websocketServer`. This function
initiates a server on the specified port, and returns a handle for listening
and closing the server.

The server no longer automatically closes when clients disconnect - the server
must close itself through the handle.

Clients are detected through the websocket_server_connect event, which
the server handle's `listen` function wraps around. This event contains the port
of the server that was connected, and the handle for this connection. The handle
has the same methods as a normal handle, but also contains a `clientID` field
that uniquely identifies this connection/client.

When a server connection receives a message, it sends the `websocket_server_message`
event, which consists of the client ID, the message, and whether the message
is a binary message. Likewise, when the connection is closed, the
`websocket_server_closed` message is sent with the client ID, and optionally the
code the connection was closed with.