Moved cluster allocator out of Lua
This is likely the first in a series of allocation optimization attempts.
Allocation of ropes and substrings has been delegated to CraftOS-PC itself
to decouple it from the garbage collector. Hopefully this helps make these
objects more transparent to the GC, reducing the possibility of wild thrashing
caused by the huge memory use differences from allocating whole clusters.